requirements.txt
# This file may be used to create an environment using:
# $ conda create --name bdat1007 --file requirements.txt
# platform: osx-64
beautifulsoup4=4.6.0=py36_1
blas=1.0=mkl
brotlipy=0.7.0=py36h9ed2024_1003
ca-certificates=2020.12.5=h033912b_0
certifi=2020.12.5=py36h79c6626_1
cffi=1.14.5=py36h2125817_0
chardet=3.0.4=py36_1003
click=7.1.2=pyh9f0ad1d_0
cryptography=3.4.7=py36h2fd3fbb_0
dnspython=2.0.0=py_0
flask=1.1.2=py_0
idna=2.10=py_0
intel-openmp=2020.2=258
itsdangerous=1.1.0=py36_0
jinja2=2.11.2=py_0
joblib=1.0.1=pyhd8ed1ab_0
libcxx=10.0.0=1
libffi=3.3=hb1e8313_2
markupsafe=1.1.1=py36h1de35cc_0
mkl=2019.4=233
mkl-service=2.3.0=py36hfbe908c_0
mkl_fft=1.2.0=py36hc64f4ea_0
mkl_random=1.1.1=py36h959d312_0
ncurses=6.2=h0a44026_1
nltk=3.6.1=pyhd8ed1ab_0
numpy=1.19.1=py36h3b9f5b6_0
numpy-base=1.19.1=py36hcfb5961_0
openssl=1.1.1k=h0d85af4_0
pandas=1.1.3=py36hb1e8313_0
pip=21.0.1=py36hecd8cb5_0
pycparser=2.20=py_2
pymongo=3.9.0=py36h0a44026_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pysocks=1.7.1=py36hecd8cb5_0
python=3.6.13=h88f2d9e_0
python-dateutil=2.8.1=py_0
python-dotenv=0.17.0=pyhd8ed1ab_0
python_abi=3.6=1_cp36m
pytz=2020.1=py_0
readline=8.1=h9ed2024_0
regex=2021.4.4=py36hfa26744_0
requests=2.25.1=pyhd3eb1b0_0
selenium=3.14.0=py36h1de35cc_0
setuptools=52.0.0=py36hecd8cb5_0
six=1.15.0=py36hecd8cb5_0
sqlite=3.35.4=hce871da_0
termcolor=1.1.0=py36hecd8cb5_1
textblob=0.15.3=py_0
tk=8.6.10=hb0a8c7a_0
tqdm=4.60.0=pyhd8ed1ab_0
urllib3=1.26.4=pyhd3eb1b0_0
vadersentiment=3.3.2=pyhd8ed1ab_0
werkzeug=1.0.1=py_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
zlib=1.2.11=h1de35cc_3
